% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_cell_tower_sdf.R
\docType{data}
\name{ms_cell_tower_sdf}
\alias{ms_cell_tower_sdf}
\title{Mississippi Cell Towers}
\format{
SF tibble. 582 X 20 POINT.
}
\source{
TBD
}
\usage{
ms_cell_tower_sdf
}
\description{
Mississippi cell tower data (2018) .
}
\details{
TBD
}
\keyword{datasets}
