% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_counties_tigris_sdf.R
\docType{data}
\name{ms_counties_tigris_sdf}
\alias{ms_counties_tigris_sdf}
\title{Mississippi Counties}
\format{
SF tibble. 82 X 18 MULTIPOLYGON
}
\source{
U.S. Census TIGER Lines
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}
}
\usage{
ms_counties_tigris_sdf
}
\description{
Mississippi Counties boundaries from U.S. Census TIGER database.
Used for subsetting NLD, NID, NBI and other features from nation-wide layers,
or for subsetting MS-wide layers to individual MS counties.
}
\details{
The official NLD data catalog is at \url{https://levees.sec.usace.army.mil/data-services/dictionary/}.
A Shinylive data catalog app is availble at \url{https://virtual-inertia-nld-data-catalog.netlify.app/}.
Consider readr::type_convert() or dplyr::mutate_at() or mutate_if() for char column
conversion to number or integer.
All NLD WFS layers are nationwide. Mississippi data were subset using intersection
with the U.S. Census TIGER State boundary for Mississippi (2023).
}
\keyword{datasets}
