% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fire_stations_sdf.R
\docType{data}
\name{ms_fire_stations_sdf}
\alias{ms_fire_stations_sdf}
\title{Mississippi Fire Stations}
\format{
SF tibble. 1,043 X 15 POINT.
}
\source{
TBD
}
\usage{
ms_fire_stations_sdf
}
\description{
Mississippi Fire Department Station data.
}
\details{
TBD
}
\keyword{datasets}
