% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_highway_patrol_disricts_sdf.R
\docType{data}
\name{ms_highway_patrol_disricts_sdf}
\alias{ms_highway_patrol_disricts_sdf}
\title{Mississippi Highway Patrol Disricts}
\format{
SF tibble. 9 X 3 POLYGON.
}
\source{
TBD
}
\usage{
ms_highway_patrol_disricts_sdf
}
\description{
Mississippi Highway Patrol Disricts data .
}
\details{
TBD
}
\keyword{datasets}
