% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_hospitals_sdf.R
\docType{data}
\name{ms_hospitals_sdf}
\alias{ms_hospitals_sdf}
\title{Mississippi Hospital Facility Locations}
\format{
SF tibble. 132 X 33 POINT.
}
\source{
TBD
}
\usage{
ms_hospitals_sdf
}
\description{
Mississippi Hospital Facility Locations data .
}
\details{
TBD
}
\keyword{datasets}
