% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_hrsa_sdf.R
\docType{data}
\name{ms_hrsa_sdf}
\alias{ms_hrsa_sdf}
\title{Mississippi Hospital Response Service Areas}
\format{
SF tibble. 106 X 4 MULTIPOLYGON.
}
\source{
TBD
}
\usage{
ms_hrsa_sdf
}
\description{
Mississippi Hospital Response Service Area data .
}
\details{
TBD
}
\keyword{datasets}
