% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_lidar_sdf.R
\docType{data}
\name{ms_lidar_sdf}
\alias{ms_lidar_sdf}
\title{Mississippi Lidar Point Cloud Extents}
\format{
SF tibble. 5 POLYGON MBRs in Warren County.
}
\source{
Included
}
\usage{
ms_lidar_sdf
}
\description{
Mississippi Lidar Point Cloud Extents
for Natchez Trace (2016) .
}
\details{
Non-spatial columns contain USGS 3DEP metadata
}
\keyword{datasets}
