% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_nbi_bridges_sdf.R
\docType{data}
\name{ms_nbi_bridges_sdf}
\alias{ms_nbi_bridges_sdf}
\title{Mississippi Bridges}
\format{
SF tibble. 16,755 X 126 POINT.
}
\source{
\url{https://www.fhwa.dot.gov/bridge/nbi/ascii.cfm}.
The Coding Guide (Data Catalog) is available at
\url{https://www.fhwa.dot.gov/bridge/mtguide.pdf}.
}
\usage{
ms_nbi_bridges_sdf
}
\description{
Mississippi Bridges data from the National Bridge Inventory (NBI).
This includes highway overpasses,
which is why there are over 16,000 in one State alone.
}
\details{
Available in State specific CSV files.
Programmatically converted uniquely formatted latitudes and longitudes
into standard DDD.DDDDDD format
and then programmatically generated ST_POINT column.
}
\keyword{datasets}
