% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_nid_dams_sdf.R
\docType{data}
\name{ms_nid_dams_sdf}
\alias{ms_nid_dams_sdf}
\title{Mississippi Dams}
\format{
SF tibble. 6,093 X 86 POINT.
}
\source{
\url{https://nid.sec.usace.army.mil/#/downloads}
}
\usage{
ms_nid_dams_sdf
}
\description{
Mississippi Dams data from the National Inventory of Dams (NID).
}
\details{
Subset from the Nation using TIGER MS State Boundary.
}
\keyword{datasets}
