% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_police_51_fields_sdf.R
\docType{data}
\name{ms_police_51_fields_sdf}
\alias{ms_police_51_fields_sdf}
\title{Mississippi Police Station Locations}
\format{
SF tibble. 431 X 51 POINT.
}
\source{
TBD
}
\usage{
ms_police_51_fields_sdf
}
\description{
Mississippi Police Station Location data .
}
\details{
TBD
}
\keyword{datasets}
