% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_public_school_districts_sdf.R
\docType{data}
\name{ms_public_school_districts_sdf}
\alias{ms_public_school_districts_sdf}
\title{Mississippi Public School Districts}
\format{
SF tibble. 138 X 15 MULTIPOLYGON.
}
\source{
TBD
}
\usage{
ms_public_school_districts_sdf
}
\description{
Mississippi Public School Districts data .
}
\details{
TBD
}
\keyword{datasets}
