% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_public_school_locations_sdf.R
\docType{data}
\name{ms_public_school_locations_sdf}
\alias{ms_public_school_locations_sdf}
\title{Mississippi Public School Facility Locations}
\format{
SF tibble. 1,062 X 27 POINT.
}
\source{
TBD
}
\usage{
ms_public_school_locations_sdf
}
\description{
Mississippi Public School Facility Location data .
}
\details{
TBD
}
\keyword{datasets}
