% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_systems_segments_sdf.R
\docType{data}
\name{ms_systems_segments_sdf}
\alias{ms_systems_segments_sdf}
\title{Mississippi System and Segment for one segment}
\format{
SF tibble 1 X 3 MULTIPOLYGON.
}
\source{
TBD
}
\usage{
ms_systems_segments_sdf
}
\description{
Mississippi System and Segment data.
leveedId   \if{html}{\out{<int>}} 1406000027
fcSystemId \if{html}{\out{<int>}} 1405000173
}
\details{
Result of a join test.
}
\keyword{datasets}
