% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_usgs_nwis_gage_sites_sdf.R
\docType{data}
\name{ms_usgs_nwis_gage_sites_sdf}
\alias{ms_usgs_nwis_gage_sites_sdf}
\title{Mississippi USGS NWIS Gage Sites}
\format{
SF tibble. 141 X 8 POINT.
}
\source{
\url{https://waterdata.usgs.gov/ms/nwis/current/?type=flow}
}
\usage{
ms_usgs_nwis_gage_sites_sdf
}
\description{
Mississippi Naional Water Information System Flow Gage Sites .
}
\details{
TBD
}
\keyword{datasets}
