% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nid_dams_sdf.R
\docType{data}
\name{nid_dams_sdf}
\alias{nid_dams_sdf}
\title{USA Dams}
\format{
SF tibble. 91,687 X 98 POINT.
}
\source{
\url{https://nid.sec.usace.army.mil/#/downloads}.
Available in both CSV and Geopackage formats.
}
\usage{
nid_dams_sdf
}
\description{
USA Dam data from the National Inventory of Dams .
Not subset for Mississippi.
}
\details{
TBD
}
\keyword{datasets}
