% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nld_domain_decodes_df.R
\docType{data}
\name{nld_domain_decodes_df}
\alias{nld_domain_decodes_df}
\title{National Levee Database domain decodes}
\format{
A tibble with 273 rows and 3 variables:
\describe{
\item{domain}{chr domain for the enumerated values}
\item{code}{num enum value (integer)}
\item{description}{chr enum description}
}
}
\source{
\url{https://levees.sec.usace.army.mil/data-services/dictionary/}
}
\usage{
nld_domain_decodes_df
}
\description{
See Data Catalog Shinylive app at \url{https://jade-blancmange-b524a1.netlify.app/}.
}
\keyword{datasets}
