% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nld_feature_catalog_df.R
\docType{data}
\name{nld_feature_catalog_df}
\alias{nld_feature_catalog_df}
\title{National Levee Database Feature Catalog}
\format{
A tibble with 206 rows and 7 variables:
\describe{
\item{field_name}{chr database field name}
\item{data_type}{chr database data type}
\item{domain}{chr domain (enum) name, if an enum}
\item{length}{num database string length, if string}
\item{definition}{database field definition}
\item{feature_name}{chr name of spatial feature class}
\item{display_order}{int display order for picklist}
}
}
\source{
\url{https://levees.sec.usace.army.mil/data-services/dictionary/}
}
\usage{
nld_feature_catalog_df
}
\description{
See Data Catalog Shinylive app at \url{https://jade-blancmange-b524a1.netlify.app/}.
}
\keyword{datasets}
