% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nld_feature_definitions_df.R
\docType{data}
\name{nld_feature_definitions_df}
\alias{nld_feature_definitions_df}
\title{National Levee Database Feature Definitions}
\format{
A tibble with 17 rows and 3 variables:
\describe{
\item{feature}{chr geometry feature class (embankment, floodwall, etc)}
\item{geometry}{chr spatial type (point, line, linestring, polygon)}
\item{definition}{chr feature definition}
}
}
\source{
\url{https://levees.sec.usace.army.mil/data-services/dictionary/}
}
\usage{
nld_feature_definitions_df
}
\description{
See Data Catalog Shinylive app at \url{https://jade-blancmange-b524a1.netlify.app/}.
}
\keyword{datasets}
