% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_state_bbox_sdf.R
\docType{data}
\name{us_state_bbox_sdf}
\alias{us_state_bbox_sdf}
\title{State Bounding Boxes}
\format{
tibble. 56 X 7
}
\source{
TBD
}
\usage{
us_state_bbox_sdf
}
\description{
Sate Bounding Boxes for convenient clipping in mapping applications .
}
\details{
xmin, xmax, ymin, ymax, STAEFP, STUSPS
}
\keyword{datasets}
