% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_state_centroid_df.R
\docType{data}
\name{us_state_centroid_df}
\alias{us_state_centroid_df}
\title{Data of centroids and zoom levels for U.S. States}
\format{
A tibble with 51 rows and 4 variables:
\describe{
\item{name}{chr State name}
\item{lat}{num Latitude of centroid in decimal degrees}
\item{lon}{num Longitude of centroid in decimal degrees}
\item{zoom}{int Zoom level for common map applications (leaflet etc)}
}
}
\source{
\url{https://developers.google.com/public-data/docs/canonical/states_csv}
}
\usage{
us_state_centroid_df
}
\description{
Latitudes and Longitudes are in decimal degrees.
}
\keyword{datasets}
